test code refactoring
